Add files via upload
Adding ability to rename and delete as templates.

+++ views/core.tpl
@@ -1,19 +1,21 @@
 <!DOCTYPE html>
 <html>
 <head>
-%pagename = pagename.replace('_',' ')
-%upcasename = str.title(pagename)
+%pagenameSpaces = pagename.replace('_',' ')
+%upcasename = str.title(pagenameSpaces)
 %bodytext = str.strip(bodytext)
 <title>{{upcasename}}</title>
 <link rel="stylesheet" type="text/css" href="/styles/dark-main.css" />
 </head>
-
 <body>
+
+<!--Menu items / top and side bar(s) -->
+
 <div class="sideBar" style="display:none" id="mySidebar">
   <button class="barButton"
   onclick="sideBar_close()">&times;</button>
   <a href="/wiki/main" class="sideBarItem">Home</a>
-  <a href="#" class="sideBarItem">About</a>
+  <a href="/wiki/about" class="sideBarItem">About</a>
   <a href="#" class="sideBarItem">Contact</a>
 </div>
 
@@ -26,6 +28,8 @@
     </div>
   </div>
 
+<!-- Page content unique to this page -->
+
   <div class="header">
     <div class="container">
       <h2>{{upcasename}}</h2>
@@ -37,23 +41,34 @@
 
 </div>
 
+<!-- Footer -->
+
 <div class="articlefoot">
-  <a href="/delete/{{pagename}}" class="footerbox">Delete</a>
-  <a href="/rename/{{pagename}}" class="footerbox">Rename</a>
-  <a href="/edit/existing/{{pagename}}" class="footerbox">Edit</a>
+
+  %if pagename not in ['main','about']:
+      <a href="/delete/{{pagename}}" class="footerbox" onclick="return confirm('Are you sure you want to delete {{pagename}}?')">Delete</a>
+      <a href="/rename/existing/{{pagename}}" class="footerbox">Rename</a>
+      <a href="/edit/existing/{{pagename}}" class="footerbox">Edit</a>
+
+
+  %else:
+      <a href="#" class="footerbox">Delete</a>
+      <a href="#" class="footerbox">Rename</a>
+      <a href="#" class="footerbox">Edit</a>
+  %end
 </div>
 
+<!-- Scripts -->
 <script>
-function sideBar_open() {
-  document.getElementById("main").style.marginLeft = "15%";
-  document.getElementById("mySidebar").style.width = "15%";
-  document.getElementById("mySidebar").style.display = "block";
-}
-function sideBar_close() {
-  document.getElementById("main").style.marginLeft = "0%";
-  document.getElementById("mySidebar").style.display = "none";
-}
+  function sideBar_open() {
+    document.getElementById("main").style.marginLeft = "15%";
+    document.getElementById("mySidebar").style.width = "15%";
+    document.getElementById("mySidebar").style.display = "block";
+  }
+  function sideBar_close() {
+    document.getElementById("main").style.marginLeft = "0%";
+    document.getElementById("mySidebar").style.display = "none";
+  }
 </script>
-
 </body>
 </html>
